4.22

+++ part4/blogilista/controllers/blogs.js
@@ -2,49 +2,43 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
+const { tokenExtractor, requestLogger, userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog
-    .find({}).populate('user', { username: 1, name: 1 })
-  
-  response.json(blogs.map(blog => blog.toJSON()))
+  try {
+    const blogs = await Blog
+      .find({}).populate('user', { username: 1, name: 1 })
+
+    response.json(blogs.map(blog => blog.toJSON()))
+  } catch (e) {
+    console.log('err', e)
+  }
 })
 
-blogsRouter.get('/:id', async (request, response, next) => {
+blogsRouter.get('/:id',  async (request, response, next) => {
   try {
     const blog = await Blog.findById(request.params.id)
     response.json(blog)
-  } catch(e) {
+  } catch (e) {
     next(e)
   }
 })
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-    return authorization.substring(7)
-  }
-  return null
-}
-
-blogsRouter.post('/', async (request, response, next) => {
-  const body = request.body
-  const token = getTokenFrom(request)
-  const decodedToken = jwt.verify(token, process.env.SECRET)
-  if (!token || !decodedToken.id) {
-    return response.status(401).json({ error: 'token missing or invalid' })
-  }
-  const user = await User.findById(decodedToken.id)
-
-/*   const user = await User.findById(body.userId) */
-  const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    user: user._id,
-    url: body.url,
-    likes: body.likes,
-  })
+blogsRouter.post('/', userExtractor, async (request, response, next) => {
   try {
+    const body = request.body
+    const user = request.user
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      user: user.id,
+      url: body.url,
+      likes: body.likes,
+    })
     const savedBlog = await blog.save()
     user.blogs = user.blogs.concat(savedBlog._id)
     await user.save()
@@ -54,11 +48,31 @@ blogsRouter.post('/', async (request, response, next) => {
   }
 })
 
-blogsRouter.delete('/:id', async (request, response, next) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response, next) => {
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
-  }catch (e) {
+    const user = request.user
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    
+    blog = await Blog.findById(request.params.id)
+    if ( blog.user.toString() === user.id.toString() ){
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    } else {
+      response.status(401).end()
+    }
+  } catch (e) {
+    next(e)
+  }
+})
+
+blogsRouter.delete('/', async (request, response, next) => {
+  try {
+    await Blog.deleteMany({})
+    response.status(200).end()
+  } catch (e) {
     next(e)
   }
 })
@@ -76,9 +90,9 @@ blogsRouter.put('/:id', async (request, response, next) => {
   try {
     const blog = await Blog.findByIdAndUpdate(request.params.id, newBlog, { new: true })
     response.json(blog)
-  } catch(e){
+  } catch (e) {
     next(e)
-  }  
+  }
 })
